refactor: Refactor folder paths for user sign-up, sign-in, and user info components
