+++ app/business/services/lecture/search-lecture.query.ts
