+++ app/business/services/lecture/taken-lecture.query.ts
