+++ app/business/services/result/result.query.ts
