+++ app/business/services/result/result.type.ts
