+++ app/ui/lecture/upload-taken-lecture/upload-taken-lecture.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { registerUserGrade } from '@/app/business/service/lecture/taken-lecture.command';
+import { registerUserGrade } from '@/app/business/services/lecture/taken-lecture.command';
 import UploadPdf from '@/app/ui/view/molecule/upload-pdf/upload-pdf';
 import Form from '../../view/molecule/form';
