+++ app/ui/result/result-category-detail-lecture/result-cagegory-detail-lecture.tsx
@@ -2,7 +2,7 @@ import { Table } from '../../view/molecule/table';
 import AnnounceMessageBox from '@/app/ui/view/molecule/announce-message-box/announce-massage-box';
 import LabelContainer from '@/app/ui/view/atom/label-container/label-container';
 import { cn } from '@/app/utils/shadcn/utils';
-import { ResultCategoryDetailLectures } from '@/app/business/service/result/result.type';
+import { ResultCategoryDetailLectures } from '@/app/business/services/result/result.type';
 
 const headerInfo = ['과목코드', '과목명', '학점'];
